Cursor Controls
Adjusted corner and translate calculations so that they work. I also restructured the translation of the cursor so it works.

I also added a 2 button inputAction for controlling the cursor and a 1 button inputAction for accelerating the cursor while held.

There is now a bug with negative position values ( I think)