Fiery Raze
Added a second spell - Fiery Raze.
In doing so, I added the Scaling function to SpellsManager, which should be usable by any spell type (although it will take a tiny bit of support in the movement script possibly)

This also has added a new spawning location (IIRC I called it AdjacentCorners) for spells which may be useful with or without some tweaking.

There is still no controls for spells set up yet, just one inputAction which I manually put a spell name string into in the CharacterControls script

That should be all but there may be other changes.