ScriptableSpell editor and fixes
Finished the custom editor for ScriptableSpllObjects and added all variables.
Fixed the existing spells to function (hasty jolt requires a hitbox after the scriptableSpell UI rework though)

Fixed variable control inputs - clamped triggers in the action maps so it will always trigger the same spell. It may eventually be necessary to clamp all inputs in SpellcastingSelection to make it so nothing breaks on rebind.