Spell Position Sync Online
Added a check to make sure the positions of spells stay synced up, including cursor position on cast. This includes restructuring a lot of the code for the cursor, though it was a mess before (and still is for some parts) and so it can't be a bad thing to put some time into it.