Re-implemented local multiplayer + networking changes
For the networking changes, I changed the game from a 10hz tick rate online to 50 Hz, which inevitably caused a few major bugs which took a few hours to fix. It required me to fix a bunch of code, but it all works now and is written much better and should be easier to change in the future.