Added a system for status effects
Currently the only status effect is one for the Fire Dash spell that stops the player from moving, but I'll add plenty more in the future. I also started work on an event system for when health is changed. Once that's working I'll probably create a generic Stat class that includes a max value, current value, and automatically calls an event when it changes.

I'm also trying to figure out some sort of predictive spawning stuff so that the spells sync better online. Either I'm gonna have to migrate to something like Netcode for Entities instead of Netcode for GameObjects, or I'll have to spend a while figuring out how to set up some very basic predictive spawning through NGO.